Add files via upload
The Project is the idea we pitched on NIC BlockChain Hackathon organized by TrustWork Community with collaboration of other major and well known organizations and communities. 

The Idea of making a decentralized Blood bank where everyone has access to all the information of available blood types.  location and  procedure to get the blood .The Idea inforce the maximum protection from blood mafia, an ambiguite behavior is detected with ease.

Please start the reporsity if you find the idea helpful