Google Geocode API Project.
Date : Jan 26, 2019

The Project consist of ES6 basics including arrow functions, fetch api method that returns a promise as well as for-of loop. 
User is allowed to enter any location and the application will respond with the accurate geographical location including latitude and longitude. 

This is a practice project to get my hands wet before jumping into JS frameworks